chore:Condensing the latex test backend into multiple files
