+++ docling/backend/html_backend.py
@@ -559,7 +559,7 @@ class HTMLDocumentBackend(DeclarativeDocumentBackend):
             return
 
         try:
-            from playwright.sync_api import sync_playwright
+            from playwright.sync_api import sync_playwright  # type: ignore
         except ImportError as exc:
             raise RuntimeError(
                 "Playwright is required for HTML rendering. "
